Attempt to solve linting performance issues on pages with many code blocks using a web worker
The problem is, due to Chromium bug https://crbug.com/357664 , you can't load a web worker
from a content script normally; you have to bring in a library some guy wrote which patches the
Worker class and loads a whole bunch of stuff hackily to fix it. It still is only working ~80% of the time
due to a race condition.